chore: 🤖 add config for middleware
